pyproject.toml
[project]
name = "csv-pipeline"
version = "0.1.0"
description = "CSV → SQLite → Text-to-SQL pipeline"
requires-python = ">=3.12"
dependencies = [
  "anthropic>=0.50.0",
  "google-genai>=1.0.0",
  "openai>=1.80.0",
  "pandas>=2.2.0",
  "python-dotenv>=1.0.0",
  "chardet>=5.0.0",
  "tqdm>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
